Update to Bombe 0.4.0
