+++ changelogs/0.6.0.md
@@ -36,6 +36,11 @@ def mappings = MappingSetDsl.create {
 To use the Groovy DSL, simply add `org.cadixdev:lorenz-dsl-groovy:0.6.0` to your build
 tool.
 
+## ASM Module
+
+Inline with changes made in Bombe 0.4.0, the ASM package is no more - with both classes
+now available under the same package in the core Lorenz module.
+
 ## Conveniences
 
 - `Mapping#set(key, value)` now returns the `Mapping` allowing for fluent method calls.
