+++ gradle.properties
@@ -6,6 +6,7 @@ inceptionYear = 2016
 # Build Settings
 javaVersion = 1.8
 groovyVersion = 2.5.8
-bombeVersion = 0.4.0-SNAPSHOT
+asmVersion = 7.1
+bombeVersion = 0.4.0
 junitVersion = 5.5.1
 spockVersion = 1.3-groovy-2.5
