--- lorenz-asm/gradle.properties
@@ -1,4 +1,0 @@
-name = Lorenz-ASM
-description = Utility classes for using Lorenz with ASM.
-url = https://www.jamiemansfield.me/projects/lorenz
-inceptionYear = 2018
