+++ lorenz/src/test/java/org/cadixdev/lorenz/test/asm/LorenzRemapperTest.java
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  */
 
-package org.cadixdev.lorenz.asm.test;
+package org.cadixdev.lorenz.test.asm;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
