+++ lorenz/src/test/java/test/inheritance/TestClass.java
