+++ lorenz/src/test/java/test/inheritance/a/BaseClass.java
