+++ lorenz/src/test/java/test/inheritance/a/PublicBaseClass.java
